Frontend is now working
Fixed the path issues and made a home page
