+++ apps/humanoid-ai-ui/index.html
@@ -7,5 +7,6 @@
 </head>
 <body>
   <div id="root"></div>
+  <script type="module" src="/src/main.tsx"></script>
 </body>
 </html>
