+++ apps/humanoid-ai-ui/src/app/app.css
@@ -1,0 +1,9 @@
+body {
+  font-family: Arial, sans-serif;
+  margin: 0;
+  padding: 0;
+}
+
+.App {
+  text-align: center;
+}
