+++ apps/humanoid-ai-ui/src/app/app.tsx
@@ -1,48 +1,11 @@
-import NxWelcome from './nx-welcome';
+import React from 'react';
+import Home from '../components/Home';
+import './app.css';
 
-import { Route, Routes, Link } from 'react-router-dom';
-
-export function App() {
+function App() {
   return (
-    <div>
-      <NxWelcome title="humanoid-ai-ui" />
-
-      {/* START: routes */}
-      {/* These routes and navigation have been generated for you */}
-      {/* Feel free to move and update them to fit your needs */}
-      <br />
-      <hr />
-      <br />
-      <div role="navigation">
-        <ul>
-          <li>
-            <Link to="/">Home</Link>
-          </li>
-          <li>
-            <Link to="/page-2">Page 2</Link>
-          </li>
-        </ul>
-      </div>
-      <Routes>
-        <Route
-          path="/"
-          element={
-            <div>
-              This is the generated root route.{' '}
-              <Link to="/page-2">Click here for page 2.</Link>
-            </div>
-          }
-        />
-        <Route
-          path="/page-2"
-          element={
-            <div>
-              <Link to="/">Click here to go back to root page.</Link>
-            </div>
-          }
-        />
-      </Routes>
-      {/* END: routes */}
+    <div className="App">
+      <Home />
     </div>
   );
 }
