+++ apps/humanoid-ai-ui/src/main.tsx
@@ -1,7 +1,6 @@
 import { StrictMode } from 'react';
 import * as ReactDOM from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
-
 import App from './app/app';
 
 const root = ReactDOM.createRoot(
